+++ lib/wuManber.ml
@@ -7,25 +7,15 @@ module I = Int63
 module BitOps = struct
   (** A collection of bitwise operations used in the Wu and Manber algorithm. *)
 
-  (** [shift_or ~mismatch bv] takes a bitvector [bv] and returns a new bitvector
-      by doing a left shift of [bv] and then doing a logical or with
-      [~mismatch]. See the {!Matcher} module for a description of mismatch
-      bitvectors. *)
-  let shift_or ~mismatch x : I.t =
-    I.lshift1 x
-    |> I.logor mismatch
+  let[@inline] match_error ~pattern_length (bvs : I.t array) : int option =
+    Array.find_index_opt (I.bit_is_zero ~n:(pattern_length - 1)) bvs
 
-  let match_error ~pattern_length (bvs : I.t array) : int option =
-    let idx = pattern_length - 1 in
-    Array.find_index_opt (I.bit_is_zero ~n:idx) bvs
-
-  let is_match ~pattern_length (bvs : I.t array) : bool =
-    match_error ~pattern_length bvs
+  let[@inline] is_match ~pattern_length (bvs : I.t array) : bool =
+    (match_error[@inlined]) ~pattern_length bvs
     |> Option.is_some
 end
 
 module WuManber = struct
-  open BitOps
   let initial_bvs ~k : I.t array =
     let arr = Array.make (k+1) I.minus_one in
     for i = 1 to k do
@@ -33,19 +23,23 @@ module WuManber = struct
     done;
     arr
 
+  let[@inline] transition1_so ~shift_or1 ~input0 ~output0 : I.t =
+    let open I.Infix in
+    shift_or1 && (I.lshift1 (input0 && output0)) && input0
+
   let next_bvs ~mismatch (input : I.t array) : I.t array =
     (* Self Transitions, i.e. R2/D2/S2 *)
     let len = Array.length input in
-    let output = Array.map (shift_or ~mismatch) input in
+    let output = Array.map (I.shift_or ~mismatch) input in
     (* inserts, shifts, deletes *)
-    for i = 1 to (len - 1) do
-      output.(i) <-
-        let rjd1 = input.(i-1) in
-        let rj1d1 = output.(i-1) in
-        let open I.Infix in
-        (* self-trans & shift (substitution and delete) & insert *)
-        output.(i) && (I.lshift1 (rjd1 && rj1d1)) && rjd1
-    done;
+    let () =
+      for i = 1 to (len - 1) do
+        output.(i) <-
+          let rjd1 = input.(i-1) in
+          let rj1d1 = output.(i-1) in
+          (transition1_so[@inlined]) ~shift_or1:output.(i) ~input0:rjd1 ~output0:rj1d1
+      done
+    in
     output
 end
 
@@ -68,7 +62,7 @@ module RightLeaningWuManber = struct
           let open I.Infix in
           (shifted && del_mask)
         in
-        let del_mask = I.lshift1 @@ del_and_shifted in
+        let del_mask = I.lshift1 del_and_shifted in
         let new_output_i =
           let open I.Infix in
           (* self-trans && delete & shift (substitution) & insert *)
@@ -86,4 +80,5 @@ module RightLeaningWuManber = struct
 
   let feed_sentinel (input : I.t array) : I.t array =
     next_bvs ~mismatch:(I.minus_one) input
+
 end