+++ CHANGES.md
@@ -1,4 +1,8 @@
 # Changelog
+## [unreleased]
+### Added
+- Specialized versions of the algorithm for use in hot loops.
+
 ## [0.1.0]
 
 Initial release.